Abstraindo validacoes
