+++ src/app/controllers/SessionController.js
@@ -1,21 +1,10 @@
 import jwt from 'jsonwebtoken';
-import * as Yup from 'yup';
 import User from '../models/User';
 import File from '../models/File';
 import authConfig from '../../config/auth';
 
 class SessionController {
   async store(req, res) {
-    const schema = Yup.object().shape({
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string().required(),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return res.status(400).json({ error: 'Validation fails' });
-    }
     const { email, password } = req.body;
     const user = await User.findOne({
       where: { email },
